+++ modules/profiles/manifests/editor/gimp.pp
@@ -4,9 +4,11 @@
 class profiles::editor::gimp {
   package { 'gimp': }
 
-  exec { 'CleanupGimpDesktopShortcut':
-    command     => "Remove-Item -Path 'C:\\Users\\${lookup('camper_username')}\\Desktop\\GIMP*.lnk'",
-    refreshonly => true,
-    subscribe   => Package['gimp'],
+  if $facts['os']['family'] == 'windows' {
+    exec { 'CleanupGimpDesktopShortcut':
+      command     => "Remove-Item -Path 'C:\\Users\\${lookup('camper_username')}\\Desktop\\GIMP*.lnk'",
+      refreshonly => true,
+      subscribe   => Package['gimp'],
+    }
   }
 }
